Upgrade to Boost 1.66.0
The reason is that I ran into a C++17 compatibility issue with
`boost::numeric_cast`, namely this one:
https://github.com/boostorg/numeric_conversion/issues/5

Note that the only reason to use Boost 1.65.1 in the first place was
that it is the latest version supported by the the FindBoost script
bundled with CMake 3.9.  Now, we're using a custom FindBoost script
developed specifically for Bincrafters' modular Boost packages
(the `cmake_findboost_modular` package), so this is a non-issue for us.