Set up custom logging sink
The default logging sink in Boost.Log apparently does not support custom
log level definitions, which was the cause of issue #48. This sets up a
custom sink that writes to the standard log stream (which is really just
the standard error stream) with custom message formatting, fixing that
issue.

The nice part is that this prepares the way for other custom sinks in
the future, like log files, network logging or whatever.