Move pseudo-asynchronous slave wrapper to main lib
This was previously only available for the tests, but it is generally
useful functionality that will become necessary when we merge the C and
C++ co-simulation APIs (to wrap a `cse::fmi::slave_instance` in a
`cse::async_slave`).

I also removed its dependency on an event loop, which was unnecessary
and only there for testing purposes.