Simplify cse::slave::setup()
This removes the `slaveName` and `executionName` parameters from
`slave::setup()`.  The fact that they were there in the first place is a
holdover from Coral, and I don't believe they are needed here. With them
gone, the mapping from the `slave` interface to the FMI functions is more
straightforward, and consequently the implementation of
`fmi::vX::slave_instance` is too.  This addresses issue #62.