Consistently use relative tolerance in setup()
FMI specifies that the tolerance is "usually" relative, but since the
FMU code has no way of determining whether it is relative or absolute,
it just has to pick one. Presumably, most will assume the former, since
that is what the standard says is most common.  We can do better by
*requiring* it.