Use the "official" Boost package from conan-center
Until now, we have used the Boost packages from Bincrafters, which had
one package per sub-library.  This has led to very high compile times
due to the large number of include directories (89 at the last count!).
With this change, compile times seem to improve significantly.

Another advantage of using the official package is that we can use the
FindBoost script bundled with CMake, which has better dependency
detection, allowing us to simplify the CMake build scripts a bit.

Now, we don't need to add the Bincrafters repository as a Conan remote
either, as all packages are exposed through conan-center, so I removed
that step from the Jenkinsfile.