Set fiber stack size to 1MiB in pseudo_async_slave
The default stack size in Boost.Fiber is just a few kilobytes, which
caused problems with one particular FMU that tried to put a
several-hundred-kilobyte object on the stack.  Third-party code should
not have to care that CSE uses fibers, so we increase the stack size to
1 MiB (which is the default thread stack size in MSVC).

`pseudo_async_slave` is only an interim solution anyway.  When we start
running slaves in separate threads/processes this will be a non-issue.