Workaround to enable Linux RelWithDebInfo builds
For some reason, `libboost_unit_test_framework.so` depends on
`libboost_timer.so` when building with the `RelWithDebInfo`
configuration.  This is probably a bug in either CMake's FindBoost
script or in Conan's Boost package.  I don't think we are causing it
ourselves.  Here, I've add Boost.Timer as an explicit dependency in
order to work around the issue.