clang-format: Drop space before initializer lists
This seems more consistent with what most people are doing (it led to
less changes when running the housekeeping script), and to me it makes
the most sense.  The braces in an initializer list have the same purpose
as the parentheses in a (constructor) function call, and should be
formatted the same.