Fix #284, "fixed_step_algorithm crashes when..."
The problem was that fixed_step_algorithm::impl::disconnect_variable
didn't check whether `it` pointed to an actual element before trying to
erase it.

This also fixes an unrelated hidden bug in the same function:
The `conns` variable was a *copy* of the connection list, and not a
reference to it as was intended.  This would have rendered the entire
function a no-op.