Fix #361, "Memory error when moving a cse::uri" (#372)
I've replaced the internal `std::string_view` members in `cse::uri` with
a new type (`subrange`) which is independent of absolute memory location
but otherwise contains the same information.  To be able to take
advantage of all the nice functionality built into `std::string_view`,
`subrange` is only used for storage, not for processing.