Avoid deadlock in cse_execution_stop() (#408)
This fixes #398. The idea is to avoid the deadlock by letting other fibers run while we wait for the "simulation thread" to finish. This is achieved by wrapping the simulate_until() call in a boost::fibers::packaged_task that signals its completion to a boost::fibers::future in a way that is both fiber friendly and thread safe.