Revert "Set scm.url="auto" in conanfile.py (#503)" (#511)
This reverts commit 7d1939d14dfb70cfd437aceec182f59d196bf8ef.

The change caused a problem in the case where the recipe is
downloaded from Artifactory, but the package is built locally
(e.g. with `conan install --build=<package>`).

The problem is that `auto` gets replaced with a HTTPS repository
address, because that's what Jenkins uses.  Most people/systems
haven't set up the proper credentials for HTTPS access to GitHub,
however.

This reopens issue #502.