Use "1" as true value for environment variables (#645)
Conan only understands the values "True" or "1" for its configuration
variables.  However, in a YAML file, the value "True" gets intepreted as
a boolean rather than a string and converted to lowercase "true" at some
point by GitHub Actions. Conan does *not* understand that and interprets
it as "False" instead.

This is not actually a problem in the part of `ci-conan.yml` which
generates the `Dockerfile`, as Docker does no such interpretation.  But
it's less confusing and more future-proof to stick to one convention
everywhere.