Correct or remove minimum version specs in README (#704)
Pretty much all of them were wrong or misleading. It's very hard to remember to keep these up to date, and it's also difficult to know exactly which versions will work. So from now on, we specify just the compiler versions we're testing in CI and leave the other tool versions unspecified.